requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
aiosqlite==0.19.0
alembic==1.12.0
amqp==5.1.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blinker==1.6.2
cachelib==0.10.2
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.3
command-not-found==0.3
cryptography==39.0.1
databases==0.8.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docxcompose==1.4.0
docxtpl==0.16.7
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.2
fastapi-mail==1.4.1
fastjsonschema==2.17.1
flower==2.0.1
fqdn==1.5.1
greenlet==2.0.1
gunicorn==21.2.0
h11==0.14.0
httplib2==0.20.2
humanize==4.8.0
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==23.5.0
kombu==5.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==8.10.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.0
oauthlib==3.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pycoingecko==3.1.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.7.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.19.3
pytest==7.3.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==4.5.5
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
speaklater==1.3
SQLAlchemy==1.4.0
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.27.0
systemd-python==234
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
ulid==1.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
WTForms==2.3.3
zipp==1.0.0
